refactor(database): break circular barrel imports in queries
queries/*.js imported schema and db helpers back from the database/index.js
barrel that re-exports them, creating an index.js <-> queries cycle. After a
new main-process entry shifted Rollup's chunk boundaries, this surfaced as 30
"was reexported through module" warnings with undefined cross-chunk load order.

Move getDrizzleDb/getReadonlyDrizzleDb/executeRawQuery into manager.js (next to
getStudyDatabase they wrap), and point queries at ../models.js for schema and
../manager.js for helpers instead of the barrel. index.js still re-exports
everything, so IPC/service callers are unchanged.
