+++ src/main/database/index.js
@@ -4,7 +4,14 @@
  */
 
 import { eq, desc } from 'drizzle-orm'
-import { getStudyDatabase, closeStudyDatabase, closeAllDatabases } from './manager.js'
+import {
+  getStudyDatabase,
+  closeStudyDatabase,
+  closeAllDatabases,
+  getDrizzleDb,
+  getReadonlyDrizzleDb,
+  executeRawQuery
+} from './manager.js'
 import {
   deployments,
   media,
@@ -24,7 +31,14 @@ import log from '../services/logger.js'
 
 // Re-export schema and manager functions
 export { deployments, media, observations, modelRuns, modelOutputs, metadata, jobs }
-export { getStudyDatabase, closeStudyDatabase, closeAllDatabases }
+export {
+  getStudyDatabase,
+  closeStudyDatabase,
+  closeAllDatabases,
+  getDrizzleDb,
+  getReadonlyDrizzleDb,
+  executeRawQuery
+}
 
 // Re-export Zod validation schemas
 export {
@@ -42,50 +56,6 @@ export {
   manasRawOutputSchema,
   rawOutputSchema
 } from './validators.js'
-
-/**
- * Helper function to get Drizzle database instance for a study
- * @param {string} studyId - Study identifier
- * @param {string} dbPath - Path to database file
- * @param {Object} options - Database options (e.g., {readonly: true})
- * @returns {Promise<Object>} Drizzle database instance
- */
-export async function getDrizzleDb(studyId, dbPath, options = {}) {
-  const manager = await getStudyDatabase(studyId, dbPath, options)
-  return manager.getDb()
-}
-
-/**
- * Helper function to get a readonly Drizzle database instance for a study
- * @param {string} studyId - Study identifier
- * @param {string} dbPath - Path to database file
- * @returns {Promise<Object>} Readonly Drizzle database instance
- */
-export async function getReadonlyDrizzleDb(studyId, dbPath) {
-  const manager = await getStudyDatabase(studyId, dbPath, { readonly: true })
-  return manager.getDb()
-}
-
-/**
- * Helper function to execute raw SQL queries when needed
- * @param {string} studyId - Study identifier
- * @param {string} dbPath - Path to database file
- * @param {string} query - SQL query
- * @param {Array} params - Query parameters
- * @returns {Promise<Array>} Query results
- */
-export async function executeRawQuery(studyId, dbPath, query, params = []) {
-  const manager = await getStudyDatabase(studyId, dbPath)
-  const sqlite = manager.getSqlite()
-
-  try {
-    const statement = sqlite.prepare(query)
-    return statement.all(params)
-  } catch (error) {
-    log.error(`[DB] Raw query failed for study ${studyId}:`, error)
-    throw error
-  }
-}
 
 // ============================================================================
 // Metadata CRUD operations
