+++ src/main/database/queries/overview.js
@@ -9,7 +9,7 @@
  * against the bundled speciesInfo dictionary for the threatened tally).
  */
 
-import { getStudyDatabase } from '../index.js'
+import { getStudyDatabase } from '../manager.js'
 import log from '../../services/logger.js'
 import { getStudyIdFromPath } from './utils.js'
 import { resolveSpeciesInfo } from '../../../shared/speciesInfo/resolver.js'
